// web/src/api/models/competitorChangesPlatform.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * AppStoreCat API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type CompetitorChangesPlatform = typeof CompetitorChangesPlatform[keyof typeof CompetitorChangesPlatform];


export const CompetitorChangesPlatform = {
  ios: 'ios',
  android: 'android',
} as const;